Fixes #7, add cached flatpages results.
The issue related in #7 happened because I wasn't considering
that we can have anything in url before a FlatPage url, that's
fixed by not return 404 if part of url isn't a FlatPage.

Cached flatpage results was added to breadcrumbs.views.flatpage,
to avoid hit db to query every request and every FlatPage
in building all parts of breadcrumb.